requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-firestore==2.13.1
google-api-python-client==2.108.0
sendgrid==6.10.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

